+++ CHANGELOG.md
@@ -1,5 +1,38 @@
 # Change log for Microsoft365DSC
 
+# 1.26.603.1
+
+* AADConditionalAccessPolicy
+  * Improved handling of multiple custom security attributes instead of just one
+    in `ServicePrincipalFilterRule`.
+* AADFilteringPolicyRule
+  * Fixed an issue where an incorrect number of instances to export was shown.
+* AADOrganizationCertificateBasedAuthConfiguration
+  * Fixed an issue where the `Certificate` property was incorrectly converted using `ToBase64String` when the Graph SDK already returns it as a Base64 string. FIXES [#7193](https://github.com/Microsoft365DSC/Microsoft365DSC/issues/7193)
+* AzureRoleDefinition
+  * Initial Release
+    FIXES [#7077](https://github.com/Microsoft365DSC/Microsoft365DSC/issues/7077)
+* EXOReportSubmissionPolicy
+  * Removed export line for the configuration because it is single instance.
+* IntuneDeviceCompliancePolicyiOS
+  * Fixed an issue when omitting `ScheduledActionsForRule` during creation.
+* IntuneDeviceCompliancePolicyMacOS
+  * Fixed an issue when omitting `ScheduledActionsForRule` during creation.
+* IntuneDeviceConfigurationPolicyAndroidOpenSourceProject
+  * Fixed an issue when updating an instance of the resource.
+* IntuneRoleDefinition
+  * Fixed an issue when omitting `notAllowedResourceActions`.
+* O365OrgSettings
+  * Fixed an issue when updating the `Office Online Unused API` service principal.
+    FIXES [#7175](https://github.com/Microsoft365DSC/Microsoft365DSC/issues/7175)
+* M365DSCPermissions
+  * Fixed an issue when updating app permissions.
+    FIXES [#7182](https://github.com/Microsoft365DSC/Microsoft365DSC/issues/7182)
+* M365DSCReverse
+  * Fixed an issue where the number of resources to export could be wrong.
+* MISC
+  * Fixed an issue where `Az.Accounts` failed to import.
+
 # 1.26.527.1
 
 * AADAccessReviewDefinition